+++ Core/Utilities/Time.py
@@ -1,4 +1,3 @@
-# $HeadURL$
 """
 DIRAC Times module
 Support for basic Date and Time operations
@@ -31,6 +30,8 @@ import time as nativetime
 import datetime
 from types import StringTypes
 
+from DIRAC import gLogger
+
 # Some useful constants for time operations
 microsecond = datetime.timedelta( microseconds = 1 )
 second = datetime.timedelta( seconds = 1 )
@@ -40,6 +41,21 @@ day = datetime.timedelta( days = 1 )
 week = datetime.timedelta( days = 7 )
 
 dt = datetime.datetime( 2000, 1, 1 )
+
+def timeThis( method ):
+  """ Function to be used as a decorator for timing other functions/methods
+  """
+
+  def timed( *args, **kw ):
+    ts = time.time()
+    result = method( *args, **kw )
+    te = time.time()
+
+    gLogger.verbose( "Exec time === ", " function %r arguments len: %d -> %2.2f sec" % ( method.__name__, len( kw ), te - ts ) )
+    return result
+
+  return timed
+
 
 def dateTime():
   """
